Fixed total progress bar problems with many files in joblist

+++ include/bonkenc.h
@@ -110,7 +110,7 @@ namespace BonkEnc
 			Bool			 dontUpdateInfo;
 			Int			 encoder_activedrive;
 
-			Int			 totalSamples;
+			Int64			 totalSamples;
 			Float			 totalSamplesDone;
 
 			Int			 startTicks;
